moving response exception definition to the response lib file
No need to have a separate exceptions lib file just for these.  Plus
its better to keep the definitions near where they are used for quick
reference.
