+++ lib/and-son/response.rb
@@ -1,7 +1,5 @@
 require 'sanford-protocol'
 
-require 'and-son/exceptions'
-
 module AndSon
 
   class Response < Struct.new(:protocol_response)
@@ -40,5 +38,20 @@ module AndSon
     end
 
   end
+
+  class RequestError < RuntimeError
+    attr_reader :response
+
+    def initialize(protocol_response)
+      super(protocol_response.status.message)
+      @response = protocol_response
+    end
+  end
+
+  ClientError     = Class.new(RequestError)
+  BadRequestError = Class.new(ClientError)
+  NotFoundError   = Class.new(ClientError)
+
+  ServerError     = Class.new(RequestError)
 
 end
